Replace FileOutputStream by nio Files.newOutputStream and Channels.newOutputStream
FileOutputStream contains a finalizer which can cause excessive garbage
collector pressure when writing many TAR files.

Finalizers are only executed when the GC feels like it. Until they are
executed, objects with finalizers are not garbage collected. Internal
benchmarks have shown that hundreds of thousands of FileOutputStream
objects could remain in memory for a very long time as a result.

This patch reduces GC pressure.